Remove --install-layout=deb
We need --install-layout=deb, which is added in a patch[1] not present in Ubuntu's python3-setuptools. And GitHub hosted runners only support Ubuntu, not Debian.

Upon inspection of the aforementioned patch, I conclude that it does not implement the deb file generation itself, but merely corrects paths that we already set with --install-scripts and --prefix.

[1]: https://sources.debian.org/src/setuptools/68.1.2-2/debian/patches/install-layout.diff/